Update dependencies and eslint rules in CLI package
Updated "commander" package to version 11.0.0 and added "@commander-js/extra-typings" to support the updated commander functionality.
Also added "rxjs" and "eslint-plugin-immutable" as new dependencies, which are necessary for reactive programming and promoting immutability, respectively.
Updated the .eslintrc file to include "immutable" plugin rules. This is to ensure immutability in our codebase, a premise central to functional programming and reliable code.