Update docker-compose configuration to use custom-built image
The existing docker-compose file was using 'timberio/vector:latest-debian' image. This was changed to a custom-built image with a new Dockerfile. The motive behind changing to a custom image is to install the 'infisical' command-line interface into the image, and inject environment variables. Simultaneously, the buckets log information was updated to distinguish logs based on NODE_ID, making it easier to identify logs from specific nodes. A ._env template was added to help set required variables. Further, Dockerfile was added to facilitate the image building process.