Refactor code for querying balance and staking funds
In the process of refactoring, a new QueryManager class was created to manage all activities related to querying balances and staking funds. This was done to promote code reusability and separation of concerns. The functions that were previously in the LogStoreRegistry class have now been moved to the QueryManager. Adjustments were also made in the TokenManager and LogStoreClient to accommodate these changes.