Add LogStoreClient utility and refactor existing utility code
This commit introduced a new file `logstore-client.ts` that provides utility methods related to LogStoreClient. This removes the necessity of instantiating LogStoreClient in other parts of the code, helping to reduce code duplication and increase maintainability.

Additionally, some renaming and code moving occurred to better organize the codebase. The utility code that was in `utils.ts` was moved to its own directory `utils/`.

Lastly, a configuration setting `USE_TEST_CONFIG` was added to allow for testing configurations.