Increase test coverage and timeout in commands.test.ts
Increased overall test coverage in the commands.test.ts file for better reliability and understanding of the code. In detail, unit tests were added for 'query stake', 'store stake', 'query balance', and 'store balance' to make sure that all are functioning properly. Additionally, due to the complexity and the time taken in test execution, the TEST_TIMEOUT was updated from 20,000 to 90,000.