Add prettier-2 and dotenv packages, update jest config
Installed prettier-2 to main package and added dotenv to the client package for managing environment variables efficiently. Modified jest.config.js for both broker and client packages to integrate with prettier-2 as prettier 3 is incompatible with jest. Also added dotenv/config to setupFilesAfterEnv for injecting environment variables in testing scenarios.