Add limit to message requests in LogStore
A limit has been added to message requests in the LogStore. The new 'limit' parameter is designed to control the number of messages returned to cut down on memory use. Under the hood, it reduces the number of messages returned from each query to the Cassandra database. Additional measures have been implemented in the result stream to ensure that the total number of messages returned does not exceed the limit.