Fix key assignment in logStore plugin
This update corrects a bug in the key assignment within the logStore plugin's DataQueryFormat.ts file. Previously, the key was inaccurately assigned as 'id' instead of 'safeKey'. This error was causing issues with format retrieval in cases when 'id' was undefined. With this change, the key will correctly default to 'object' if 'safeKey' isn't presented in FORMATS.