Update and streamline integration tests in logStore plugin
The code was updated to improve the flexibility of integration tests in the logStore http-server by adding a 'format' option for raw or object data types and changing a const TRACKER_PORT from static to dynamic. Additionally, switching Http2Stream to Readable in stream handling to ensure compatibility with several Node.js versions. Plus, a new test case for raw responses correctness was added.