Update query tests and enhance stream message handling
In this update, the end-to-end query tests have been adjusted with more rigorous checks on the StreamMessage objects. This was achieved through the introduction of a more detailed object match check in the tests. Moreover, stream message handling in `HttpUtil.ts` and `Queries.ts` was enhanced.

The `MESSAGE_STORE_TIMEOUT` was also increased to provide enough buffer for retrieving messages. Additionally, the `public stream` test has been renamed to `private stream` test.