Remove ProofOfMessageStored metadata streaming mechanism from Broker
This change removes the ProofOfMessageStored checks and associated metadata streaming mechanism from the LogStoreMessageListener in the Broker.

ProofOfMessageStored was initially included to ensure the message was indeed stored by more brokers before it was acknowledged to the client. This involved a check with a threshold, a time-to-live mechanism, and some cryptographic hashing.

Now, the message gets stored immediately upon its arrival by directly calling the store method and removed all the additional checks and broadcasts associated with ProofOfMessageStored.