Remove ProofOfMessageStored related code
The files `ProofOfMessageStored.ts`, `ProofOfMessageStoredSerializerV1.ts`and the respective test file have been deleted. This has also been reflected in changes in the `exports.ts`, `SystemMessage.ts` and `index.ts` where ProofOfMessageStored was referenced. This removal was done due to ProofOfMessageStored becoming deprecated and not being used in the current application logic anymore.