Add OpenTelemetry Support for Observability
Added support of OpenTelemetry to the broker package. This will help in diagnostic tracing and querying of metrics within a distributed system which is beneficial for observing performance bottlenecks, slow network requests and other issues. It's a part of observability feature development for improving service reliability and performance monitoring.

3 new files were created under `packages/broker/src/telemetry/setup/` directory for setting up the OpenTelemetry functionality. The existing cli index file has been updated to call the setup.

New observability dependencies have been added in the package.json file and jest configuration file was updated accordingly for testing requirements.