Add observer package to Dockerfile and configure it
In this commit, the observer package is added to the Dockerfile.base from dev-network. This incorporate the observer functionality to the docker image. Necessary configuration files such as observer.json and Dockerfile.observer, a shell script start-in-docker.sh is also added to aid in starting the observer in a Docker container. This change was necessary to facilitate observer functionality in development phase.