Update config loading mechanism and development keys
A CONFIG_PATH environment variable was introduced to allow customization of the configuration file path. Now, the system will check for this variable and use the corresponding file if it's available, otherwise, it will fall back to the default path. This provides flexibility in configurations for different deployment scenarios.

Also, we updated the privateKey and network id within the development-1.env.json.