Refactor query request handling in dataQueryEndpoint
This refactoring was done for a better named responsibilities and to improve naming, modularization and encapsulation. Several files and variables were renamed for better comprehensibility - 'DataForRequest' was changed to 'QueryRequestBag', the 'get*' methods were renamed to 'getFor*' to better describe what they do. Also removed unused imports and removed logger from 'dataQueryEndpoint', as it was not utilized.