Refactor Queries and introduce LogStoreMessageStream
This commit updates the Queries.ts class to utilize the new LogStoreMessageStream instead of MessageStream. The LogStoreMessageStream class is introduced to handle the logic of processing stream messages specifically for LogStore. The diff also makes 'query' function an async and some other previously public 'last' and 'from' methods private for better encapsulation.