Implement transaction split protocol in storage
Implemented the transactionSplitProtocol to manage transaction splits in the storage of the application. This change is introduced to allow the application to handle multiple versions of storage identifiers to be encoded and decoded. The changes include exporting the transactionSplitProtocol from 'utils/storage' and replacing legacy encoding-decoding in 'ArweaveSplit' and 'ReportPoller' with the methods from the new protocol. This will make it easier to manage changes in the storage identifier protocol in the future.