Update log store to include metadata in http responses
The log store has been updated to include metadata in http responses. The code for transforming responses and formatting data has been adjusted to facilitate this. In addition, tests have also been updated to ensure that metadata is correctly included in responses. The feature can be toggled on through the 'verifyNetworkResponses' query parameter.