Add conversion output rounding and minting test
The TokenManager conversion method now includes rounding output to their respective decimals variable. This prevents decimals from appearing in conversions that should result in integers. Additionally, a test for the 'mint with usd' command has been added to the commands test suite. Another test was added to ensure the result of a conversion to wei is an integer.