Add module initialization script
The commit introduces a new script 'modules-init.sh' that changes the working directory to a specific module and runs npm install as well as executes the 'vendor-hack.sh' script. This script has been integrated into the 'modules' script in package.json, enabling a smoother and more automatic module initialization process.