Update LogStore to Observable in PropagationResolver and Dispatcher
The LogStore reference in PropagationResolver and PropagationDispatcher has been updated to an Observable, resulting in changes on how we access and use it. The rxjs package has been added to the dependencies, and necessary updates have been made in other related code files. This change aims to improve the way we manage the LogStore instance, promoting more reactive programming practices.