Update broker to remove usage of rxjs.
This commit includes changes in several files, mainly removing the dependencies on the 'rxjs' library on the broker. The 'rxjs' dependency has been removed in the pnpm-lock.yaml file and a few TypeScript files as well. The code has been refactored to avoid using 'rxjs' and its features. The changes ensure equivalent functionality without the dependency.