Add extra randomness to test id and new test case
This commit introduces an extra randomness to each test id to reduce the probability of clash. In addition, a new comprehensive test case has been added to verify the consistency of data shape, given the data is consumed in different ways. This test includes preparing stake for store manager, publishing to a stream and confirming that the content of the messages received are as expected.